--- e2e-tests-service/src/main/java/com/modsen/e2etestsservice/End2EndTestsServiceApplication.java
@@ -1,16 +1,0 @@
-package com.modsen.e2etestsservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.WebApplicationType;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication
-@EnableFeignClients
-public class End2EndTestsServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication application = new SpringApplication(End2EndTestsServiceApplication.class);
-        application.setWebApplicationType(WebApplicationType.NONE);
-        application.run(args);
-    }
-}
