--- e2e-tests-service/src/main/java/com/modsen/e2etestsservice/config/DriverClientConfig.java
@@ -1,4 +1,0 @@
-package com.modsen.e2etestsservice.config;
-
-public class DriverClientConfig {
-}
