--- e2e-tests-service/src/main/java/com/modsen/e2etestsservice/dto/request/CreateRideRequest.java
@@ -1,11 +1,0 @@
-package com.modsen.e2etestsservice.dto.request;
-
-import lombok.Builder;
-
-@Builder
-public record CreateRideRequest(
-        Long passengerId,
-        String startPoint,
-        String destinationPoint
-) {
-}
