--- e2e-tests-service/src/main/java/com/modsen/e2etestsservice/dto/request/DriverRatingRequest.java
@@ -1,10 +1,0 @@
-package com.modsen.e2etestsservice.dto.request;
-
-import lombok.Builder;
-
-@Builder
-public record DriverRatingRequest(
-        Long rideId,
-        Integer rating
-) {
-}
