--- e2e-tests-service/src/main/java/com/modsen/e2etestsservice/dto/response/PassengerResponse.java
@@ -1,11 +1,0 @@
-package com.modsen.e2etestsservice.dto.response;
-
-public record PassengerResponse(
-        Long id,
-        String firstName,
-        String lastName,
-        String email,
-        String phone,
-        Double rating
-) {
-}
