--- e2e-tests-service/src/main/java/com/modsen/e2etestsservice/dto/response/RidePageResponse.java
@@ -1,10 +1,0 @@
-package com.modsen.e2etestsservice.dto.response;
-
-import java.util.List;
-
-public record RidePageResponse(
-        List<RideResponse> rides,
-        Integer pageNumber,
-        Long total
-) {
-}
