--- e2e-tests-service/src/main/java/com/modsen/e2etestsservice/entity/DriverStatus.java
@@ -1,6 +1,0 @@
-package com.modsen.e2etestsservice.entity;
-
-public enum DriverStatus {
-    AVAILABLE,
-    UNAVAILABLE
-}
