--- e2e-tests-service/src/main/java/com/modsen/e2etestsservice/entity/RideStatus.java
@@ -1,9 +1,0 @@
-package com.modsen.e2etestsservice.entity;
-
-public enum RideStatus {
-    CREATED,
-    ACCEPTED,
-    STARTED,
-    FINISHED,
-    REJECTED
-}
