--- e2e-tests-service/src/test/java/com/modsen/e2etestsservice/config/CucumberSpringConfig.java
@@ -1,10 +1,0 @@
-package com.modsen.e2etestsservice.config;
-
-import io.cucumber.spring.CucumberContextConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@CucumberContextConfiguration
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
-public class CucumberSpringConfig {
-}
