+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/EndToEndTestServiceApplication.java
@@ -1,0 +1,16 @@
+package com.modsen.endtoendtests;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.WebApplicationType;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients
+public class EndToEndTestServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication application = new SpringApplication(EndToEndTestServiceApplication.class);
+        application.setWebApplicationType(WebApplicationType.NONE);
+        application.run(args);
+    }
+}
