+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/config/DriverClientConfig.java
@@ -1,0 +1,4 @@
+package com.modsen.endtoendtests.config;
+
+public class DriverClientConfig {
+}
