+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/dto/request/CreateRideRequest.java
@@ -1,0 +1,11 @@
+package com.modsen.endtoendtests.dto.request;
+
+import lombok.Builder;
+
+@Builder
+public record CreateRideRequest(
+        Long passengerId,
+        String startPoint,
+        String destinationPoint
+) {
+}
