+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/dto/request/PassengerRatingRequest.java
@@ -1,0 +1,10 @@
+package com.modsen.endtoendtests.dto.request;
+
+import lombok.Builder;
+
+@Builder
+public record PassengerRatingRequest(
+        Long rideId,
+        Integer rating
+) {
+}
