+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/dto/response/DriverResponse.java
@@ -1,0 +1,13 @@
+package com.modsen.endtoendtests.dto.response;
+
+import com.modsen.endtoendtests.entity.DriverStatus;
+
+public record DriverResponse(
+        String firstName,
+        String lastName,
+        String email,
+        String phone,
+        Double rating,
+        DriverStatus status
+) {
+}
