+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/dto/response/PassengerResponse.java
@@ -1,0 +1,11 @@
+package com.modsen.endtoendtests.dto.response;
+
+public record PassengerResponse(
+        Long id,
+        String firstName,
+        String lastName,
+        String email,
+        String phone,
+        Double rating
+) {
+}
