+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/dto/response/RidePageResponse.java
@@ -1,0 +1,10 @@
+package com.modsen.endtoendtests.dto.response;
+
+import java.util.List;
+
+public record RidePageResponse(
+        List<RideResponse> rides,
+        Integer pageNumber,
+        Long total
+) {
+}
