+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/dto/response/RideResponse.java
@@ -1,0 +1,44 @@
+package com.modsen.endtoendtests.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
+import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
+import com.modsen.endtoendtests.entity.RideStatus;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public record RideResponse(
+        Long id,
+        Long passengerId,
+        Long driverId,
+        String startPoint,
+        String destinationPoint,
+        RideStatus status,
+
+        @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
+        @JsonDeserialize(using = LocalDateTimeDeserializer.class)
+        @JsonSerialize(using = LocalDateTimeSerializer.class)
+        LocalDateTime createdDate,
+
+        @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
+        @JsonDeserialize(using = LocalDateTimeDeserializer.class)
+        @JsonSerialize(using = LocalDateTimeSerializer.class)
+        LocalDateTime acceptedDate,
+
+        @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
+        @JsonDeserialize(using = LocalDateTimeDeserializer.class)
+        @JsonSerialize(using = LocalDateTimeSerializer.class)
+        LocalDateTime startDate,
+
+        @JsonFormat(pattern = "yyyy-MM-dd HH:mm")
+        @JsonDeserialize(using = LocalDateTimeDeserializer.class)
+        @JsonSerialize(using = LocalDateTimeSerializer.class)
+        LocalDateTime finishDate,
+        BigDecimal estimatedCost
+) {
+}
