+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/entity/DriverStatus.java
@@ -1,0 +1,6 @@
+package com.modsen.endtoendtests.entity;
+
+public enum DriverStatus {
+    AVAILABLE,
+    UNAVAILABLE
+}
