+++ end-to-end-tests/src/main/java/com/modsen/endtoendtests/entity/RideStatus.java
@@ -1,0 +1,9 @@
+package com.modsen.endtoendtests.entity;
+
+public enum RideStatus {
+    CREATED,
+    ACCEPTED,
+    STARTED,
+    FINISHED,
+    REJECTED
+}
