+++ end-to-end-tests/src/main/resources/application.yml
@@ -1,0 +1,21 @@
+spring:
+  cloud:
+    openfeign:
+      client:
+        config:
+          passenger-service:
+            url: http://localhost:8081
+            logger-level: FULL
+          driver-service:
+            url: http://localhost:8082
+            logger-level: FULL
+          ride-service:
+            url: http://localhost:8083
+            logger-level: FULL
+          rating-service:
+            url: http://localhost:8084
+            logger-level: FULL
+
+logging:
+  level:
+    com.modsen.endtoendtests.client: DEGUG
