+++ end-to-end-tests/src/test/java/com/modsen/endtoendtests/config/CucumberEndToEndTest.java
@@ -1,0 +1,15 @@
+package com.modsen.endtoendtests.config;
+
+import org.junit.platform.suite.api.ConfigurationParameter;
+import org.junit.platform.suite.api.IncludeEngines;
+import org.junit.platform.suite.api.Suite;
+
+import static io.cucumber.junit.platform.engine.Constants.FEATURES_PROPERTY_NAME;
+import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
+
+@Suite
+@IncludeEngines("cucumber")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "com.modsen.endtoendtests")
+@ConfigurationParameter(key = FEATURES_PROPERTY_NAME, value = "src/test/resources/features/")
+public class CucumberEndToEndTest {
+}
