+++ end-to-end-tests/src/test/java/com/modsen/endtoendtests/config/CucumberSpringConfig.java
@@ -1,0 +1,9 @@
+package com.modsen.endtoendtests.config;
+
+import io.cucumber.spring.CucumberContextConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@CucumberContextConfiguration
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
+public class CucumberSpringConfig {
+}
