+++ end-to-end-tests/src/test/resources/spring.properties
