--- passenger-service/src/main/resources/application-e2e.yml
@@ -1,49 +1,0 @@
-server:
-  port: 8081
-
-spring:
-  application:
-    name: passenger-service
-
-  datasource:
-    url: jdbc:postgresql://localhost:5432/passenger_db
-    username: admin
-    password: password
-    driver-class-name: org.postgresql.Driver
-
-  jpa:
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-        format_sql: true
-    hibernate:
-      ddl-auto: validate
-    show-sql: true
-
-  liquibase:
-    change-log: db/changelog/changelog-master.xml
-
-  cloud:
-    function:
-      definition: updatePassengerRating
-    stream:
-      bindings:
-        updatePassengerRating-in-0:
-          destination: ${kafka.passenger-rating-topic-name}
-          group: ${spring.application.name}
-      kafka:
-        binder:
-          brokers:
-            - ${kafka.bootstrap-servers}
-          configuration:
-            value:
-              serializer: org.springframework.kafka.support.serializer.JsonSerializer
-            key:
-              serializer: org.apache.kafka.common.serialization.StringSerializer
-
-springdoc:
-  override-with-generic-response: false
-
-kafka:
-  bootstrap-servers: localhost:9092
-  passenger-rating-topic-name: passenger-rating-topic
